src/requirements.txt
Django==3.0.7
django-behaviors
django-environ
whitenoise
celery==4.4.7
requests==2.24.0
python-dateutil==2.8.1
gunicorn==20.0.4

djangorestframework>=3.11.0
drf-jwt
django-filter

django-healthchecks

scrapinghub==2.3.1

redis
sentry-sdk

ipython

pytest-django==3.9.0
pytest-deadfixtures
pytest-env
pytest-freezegun
pytest-mock
pytest-randomly

requests_mock
freezegun
mixer

jedi
flake8-absolute-import
flake8-bugbear
flake8-commas
flake8-django
flake8-eradicate
flake8-isort
flake8-mock
flake8-pie
flake8-print
flake8-printf-formatting
flake8-pytest
flake8-pytest-style
flake8-quotes
flake8-variables-names